refactor(git-diff): fold safety.js into the git module
Everything safety.js held was about how git gets invoked: the argument lists that
switch off the config keys naming commands, the check for the keys that cannot be
switched off, and the ref-name check for the base branch setting. That last one
was only ever called by resolveBase, a few lines further down the file it now
shares. One module for talking to git reads better than two.

index.js no longer handles git arguments at all. It used to assemble the diff
argv itself through diffArgs; it now calls diffPatch and leaves the arguments
where the rest of them live.

git.js is 321 lines and still has no npm dependencies, so both smoke suites keep
running in a checkout where the plugin has not been installed. The tests follow
the module rather than the file name: git-diff-safety.test.js is named for the
behaviour it checks, which has not changed.

No behaviour change. All three smoke suites pass, as do the end-to-end harnesses
for the trust flow, the parse cap and the layout toggle.
